+++ example/lib/main.dart
@@ -1,8 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
-import 'package:flutter_to_airplay/flutter_to_airplay.dart';
 import 'package:flutter_to_airplay_example/customIcon/custom_icon_view.dart';
 import 'package:flutter_to_airplay_example/pickFromFile/pick_from_file_view.dart';
 import 'package:flutter_to_airplay_example/pickFromURL/pick_from_url_view.dart';
@@ -11,56 +7,14 @@ void main() {
   runApp(MyApp());
 }
 
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await FlutterToAirplay.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
-
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: AppBar(
-          title: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              const Text('Flutter 2 Airplay'),
-              Text(
-                _platformVersion,
-                style: TextStyle(fontSize: 12.0),
-              ),
-            ],
-          ),
+          title: const Text('Flutter 2 Airplay'),
         ),
         body: SafeArea(
           child: Container(