+++ example/lib/pickFromFile/pick_from_file_view.dart
@@ -1,57 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_to_airplay/flutter_to_airplay.dart';
 
-class PickFromFileView extends StatefulWidget {
-  @override
-  _PickFromFileViewState createState() => _PickFromFileViewState();
-}
-
-class _PickFromFileViewState extends State<PickFromFileView> {
-  String _platformVersion = 'Unknown';
-
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await FlutterToAirplay.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
-
+class PickFromFileView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: AppBar(
-          title: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              const Text('Flutter 2 Airplay'),
-              Text(
-                _platformVersion,
-                style: TextStyle(fontSize: 12.0),
-              ),
-            ],
-          ),
+          title: const Text('Flutter 2 Airplay'),
           leading: IconButton(
             alignment: Alignment.centerLeft,
             onPressed: () => Navigator.pop(context),
@@ -61,11 +18,7 @@ class _PickFromFileViewState extends State<PickFromFileView> {
             ),
           ),
           actions: [
-            AirPlayRoutePickerView(
-              tintColor: Colors.white,
-              activeTintColor: Colors.white,
-              backgroundColor: Colors.transparent,
-            ),
+            AirPlayIconButton(),
           ],
         ),
         body: SafeArea(