+++ lib/flutter_to_airplay.dart
@@ -1,19 +1,5 @@
-export 'airplay_route_picker_view.dart';
-export 'flutter_avplayer_view.dart';
+library flutter_to_airplay;
 
-import 'dart:async';
-
-import 'package:flutter/services.dart';
-
-/// For now this class provides on method to check iOS version.
-class FlutterToAirplay {
-  // returns a method channel for given name
-  static const MethodChannel _channel =
-      const MethodChannel('flutter_to_airplay');
-
-  /// Invokes a native method [getPlatformVersion] to get iOS version.
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
-}
+export 'src/air_play_icon_button.dart';
+export 'src/airplay_route_picker_view.dart';
+export 'src/flutter_avplayer_view.dart';