+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- feat: Widget AirPlayRoutePickerView accept onShowPickerView and onClosePickerView callbacks to detect picker states
+- feat: Widget AirPlayIconButton show AirPlayRoutePickerView. This button uses your app theme and IconButton style
+
 ## 2.0.3
 
 - Update dependencies.